Get the milestone data consistently
GitHub stores PR data in both the PR and Issue data structures and the
script fetches them at different times. This means that the PR data,
which is fetched later, may lag behind. For the milestone detection this
may result a PR getting through the milestone filter (which uses the
issue data) and then getting displayed with a milestone that should be
filtered (if the milestone has been applied in that time), because right
now the filtering is on issue.milestone but the page is using
pr.milestone.

Fix that by switching the display code to issue.milestone, at list it's
consistent, either way it would render correctly on the next run.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>