Add canceled CI run backtracking
Add a check to see if the last CI twister run was cancelled and try to
backtrack up to few commits to find the last non canceled one. Should
help with github CI cancelling race conditions where the last commit is
canceled and CI runs on one of the prevoius ones.