Add DNM label highlighting
Look for for DNM labels in the dashboard and put a well visible tag on
it. This should never happen but github is weird these days so at least
show it clearly here. Not skipping these since if this condition happens
it means that CI passed (which it shouldn't) and we need to restart it
or open/close the PR or something.

Signed-off-by: Fabio Baltieri <fabiobaltieri@google.com>